+++ sdks/typescript/CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.2] - 2025-08-04
+
+### Fixed
+
+- Enabled missing `metadata` options in many places
+- Proper default behavior for SessionPeerConfig
+
+### Added
+
+- Zod validation
+- getPeerConfig in Session object
+
+### Changed
+
+- Moved parameters out of random `opts` dictionaries in many places
+- Peer and Session objects now use inner client like python SDK
+
 ## [1.2.1] - 2025-07-21
 
 ### Fixed